Move ConstraintModel class to base.py
